--- packages/meta-data-tools/src/index.ts
@@ -1,3 +1,0 @@
-import MetaDataTools, {MetaData} from './meta-data-tools';
-export {MetaData};
-export default MetaDataTools;
