--- packages/meta-data-tools/src/types.ts
@@ -1,16 +1,0 @@
-import * as React from 'react';
-
-export interface MetaDataToolsStructure {
-  metadata: Map<typeof React.Component | React.StatelessComponent<any>, MetaDataStructure>;
-  describe: (comp: typeof React.Component | React.StatelessComponent<any>) => MetaDataStructure;
-  clean: () => void;
-}
-
-export interface MetaDataStructure {
-  simulations: [Simulation];
-  addSim: (sim: Simulation) => void;
-}
-
-export interface Simulation {
-  props?: Object;
-}
