+++ packages/metadata-tools/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "meta-data-tools",
+  "name": "metadata-tools",
   "version": "1.0.0",
   "main": "cjs/index.js",
   "module": "esm/index.js",
