+++ packages/metadata-tools/src/index.ts
@@ -1,0 +1,3 @@
+import MetadataTools, {Metadata} from './metadata-tools';
+export {Metadata};
+export default MetadataTools;
