+++ packages/metadata-tools/src/tsconfig.json
