+++ packages/metadata-tools/test/tsconfig.json
